Encapsulate transaction storage filtering better
