+++ core/src/banking_stage.rs
@@ -3,7 +3,6 @@
 //! can do its processing in parallel with signature verification on the GPU.
 
 use {
-    min_max_heap::MinMaxHeap,
     crate::{
         forward_packet_batches_by_accounts::ForwardPacketBatchesByAccounts,
         immutable_deserialized_packet::ImmutableDeserializedPacket,
@@ -16,8 +15,10 @@ use {
         qos_service::QosService,
         sigverify::SigverifyTracerPacketStats,
         tracer_packet_stats::TracerPacketStats,
-        unprocessed_packet_batches::{self, *},
-        unprocessed_transaction_storage::{self, ThreadType, UnprocessedTransactionStorage},
+        unprocessed_packet_batches::*,
+        unprocessed_transaction_storage::{
+            ThreadType, UnprocessedTransactionStorage, UNPROCESSED_BUFFER_STEP_SIZE,
+        },
     },
     core::iter::repeat,
     crossbeam_channel::{
@@ -35,7 +36,6 @@ use {
     solana_perf::{
         data_budget::DataBudget,
         packet::{Packet, PacketBatch, PACKETS_PER_BATCH},
-        perf_libs,
     },
     solana_poh::poh_recorder::{BankStart, PohRecorder, PohRecorderError, TransactionRecorder},
     solana_program_runtime::timings::ExecuteTimings,
@@ -53,8 +53,8 @@ use {
     },
     solana_sdk::{
         clock::{
-            Slot, DEFAULT_TICKS_PER_SLOT, MAX_PROCESSING_AGE, MAX_TRANSACTION_FORWARDING_DELAY,
-            MAX_TRANSACTION_FORWARDING_DELAY_GPU,
+            Slot, DEFAULT_TICKS_PER_SLOT, FORWARD_TRANSACTIONS_TO_LEADER_AT_SLOT_OFFSET,
+            HOLD_TRANSACTIONS_SLOT_OFFSET, MAX_PROCESSING_AGE,
         },
         feature_set::allow_votes_to_directly_update_vote_state,
         pubkey::Pubkey,
@@ -80,10 +80,6 @@ use {
     },
 };
 
-/// Transaction forwarding
-pub const FORWARD_TRANSACTIONS_TO_LEADER_AT_SLOT_OFFSET: u64 = 2;
-pub const HOLD_TRANSACTIONS_SLOT_OFFSET: u64 = 20;
-
 // Fixed thread size seems to be fastest on GCP setup
 pub const NUM_THREADS: u32 = 6;
 
@@ -94,7 +90,6 @@ const MAX_NUM_TRANSACTIONS_PER_BATCH: usize = 64;
 const NUM_VOTE_PROCESSING_THREADS: u32 = 2;
 const MIN_THREADS_BANKING: u32 = 1;
 const MIN_TOTAL_THREADS: u32 = NUM_VOTE_PROCESSING_THREADS + MIN_THREADS_BANKING;
-pub const UNPROCESSED_BUFFER_STEP_SIZE: usize = 128;
 
 const SLOT_BOUNDARY_CHECK_PERIOD: Duration = Duration::from_millis(10);
 pub type BankingPacketBatch = (Vec<PacketBatch>, Option<SigverifyTracerPacketStats>);
@@ -996,7 +991,7 @@ impl BankingStage {
         // already processed), then add to forwarding buffer.
         let filter_forwarding_result = unprocessed_transaction_storage
             .filter_forwardable_packets_and_add_batches(
-                current_bank,
+                &current_bank,
                 &mut forward_packet_batches_by_accounts,
             );
         slot_metrics_tracker.increment_transactions_from_packets_us(
@@ -1064,294 +1059,6 @@ impl BankingStage {
             );
             unprocessed_transaction_storage.clear_forwarded_packets();
         }
-    }
-
-    /// Filter out packets that fail to sanitize, or are no longer valid (could be
-    /// too old, a duplicate of something already processed). Doing this in batches to avoid
-    /// checking bank's blockhash and status cache per transaction which could be bad for performance.
-    /// Added valid and sanitized packets to forwarding queue.
-    pub fn filter_and_forward_with_account_limits(
-        bank: &Arc<Bank>,
-        buffered_packet_batches: &mut UnprocessedPacketBatches,
-        forward_buffer: &mut ForwardPacketBatchesByAccounts,
-        batch_size: usize,
-    ) -> FilterForwardingResults {
-        let mut total_forwardable_tracer_packets: usize = 0;
-        let mut total_tracer_packets_in_buffer: usize = 0;
-        let mut total_forwardable_packets: usize = 0;
-        let mut total_packet_conversion_us: u64 = 0;
-        let mut total_filter_packets_us: u64 = 0;
-        let mut dropped_tx_before_forwarding_count: usize = 0;
-
-        let mut original_priority_queue = Self::swap_priority_queue(buffered_packet_batches);
-
-        // indicates if `forward_buffer` still accept more packets, see details at
-        // `ForwardPacketBatchesByAccounts.rs`.
-        let mut accepting_packets = true;
-        // batch iterate through buffered_packet_batches in desc priority order
-        let retained_priority_queue: MinMaxHeap<_> = original_priority_queue
-            .drain_desc()
-            .chunks(batch_size)
-            .into_iter()
-            .flat_map(|packets_to_process| {
-                let packets_to_process = packets_to_process.into_iter().collect_vec();
-
-                // Vec<bool> of same size of `packets_to_process`, each indicates
-                // corresponding packet is tracer packet.
-                let tracer_packet_indexes = packets_to_process
-                    .iter()
-                    .map(|deserialized_packet| {
-                        deserialized_packet
-                            .original_packet()
-                            .meta
-                            .is_tracer_packet()
-                    })
-                    .collect::<Vec<_>>();
-                saturating_add_assign!(
-                    total_tracer_packets_in_buffer,
-                    tracer_packet_indexes
-                        .iter()
-                        .filter(|is_tracer| **is_tracer)
-                        .count()
-                );
-
-                if accepting_packets {
-                    let (
-                        (sanitized_transactions, transaction_to_packet_indexes),
-                        packet_conversion_time,
-                    ): ((Vec<SanitizedTransaction>, Vec<usize>), _) = measure!(
-                        Self::sanitize_unforwarded_packets(
-                            buffered_packet_batches,
-                            &packets_to_process,
-                            bank,
-                        ),
-                        "sanitize_packet",
-                    );
-                    saturating_add_assign!(
-                        total_packet_conversion_us,
-                        packet_conversion_time.as_us()
-                    );
-
-                    let (forwardable_transaction_indexes, filter_packets_time) = measure!(
-                        Self::filter_invalid_transactions(&sanitized_transactions, bank,),
-                        "filter_packets",
-                    );
-                    saturating_add_assign!(total_filter_packets_us, filter_packets_time.as_us());
-
-                    for forwardable_transaction_index in &forwardable_transaction_indexes {
-                        saturating_add_assign!(total_forwardable_packets, 1);
-                        let forwardable_packet_index =
-                            transaction_to_packet_indexes[*forwardable_transaction_index];
-                        if tracer_packet_indexes[forwardable_packet_index] {
-                            saturating_add_assign!(total_forwardable_tracer_packets, 1);
-                        }
-                    }
-
-                    let accepted_packet_indexes = Self::add_filtered_packets_to_forward_buffer(
-                        forward_buffer,
-                        &packets_to_process,
-                        &sanitized_transactions,
-                        &transaction_to_packet_indexes,
-                        &forwardable_transaction_indexes,
-                        &mut dropped_tx_before_forwarding_count,
-                    );
-                    accepting_packets =
-                        accepted_packet_indexes.len() == forwardable_transaction_indexes.len();
-
-                    UnprocessedPacketBatches::mark_accepted_packets_as_forwarded(
-                        buffered_packet_batches,
-                        &packets_to_process,
-                        &accepted_packet_indexes,
-                    );
-
-                    Self::collect_retained_packets(
-                        buffered_packet_batches,
-                        &packets_to_process,
-                        &Self::prepare_filtered_packet_indexes(
-                            &transaction_to_packet_indexes,
-                            &forwardable_transaction_indexes,
-                        ),
-                    )
-                } else {
-                    // skip sanitizing and filtering if not longer able to add more packets for forwarding
-                    saturating_add_assign!(
-                        dropped_tx_before_forwarding_count,
-                        packets_to_process.len()
-                    );
-                    packets_to_process
-                }
-            })
-            .collect();
-
-        // replace packet priority queue
-        buffered_packet_batches.packet_priority_queue = retained_priority_queue;
-
-        inc_new_counter_info!(
-            "banking_stage-dropped_tx_before_forwarding",
-            dropped_tx_before_forwarding_count
-        );
-
-        FilterForwardingResults {
-            total_forwardable_packets,
-            total_tracer_packets_in_buffer,
-            total_forwardable_tracer_packets,
-            total_packet_conversion_us,
-            total_filter_packets_us,
-        }
-    }
-
-    /// Take buffered_packet_batches's priority_queue out, leave empty MinMaxHeap in its place.
-    fn swap_priority_queue(
-        buffered_packet_batches: &mut UnprocessedPacketBatches,
-    ) -> MinMaxHeap<Arc<ImmutableDeserializedPacket>> {
-        let capacity = buffered_packet_batches.capacity();
-        std::mem::replace(
-            &mut buffered_packet_batches.packet_priority_queue,
-            MinMaxHeap::with_capacity(capacity),
-        )
-    }
-
-    /// sanitize un-forwarded packet into SanitizedTransaction for validation and forwarding.
-    fn sanitize_unforwarded_packets(
-        buffered_packet_batches: &mut UnprocessedPacketBatches,
-        packets_to_process: &[Arc<ImmutableDeserializedPacket>],
-        bank: &Arc<Bank>,
-    ) -> (Vec<SanitizedTransaction>, Vec<usize>) {
-        // Get ref of ImmutableDeserializedPacket
-        let deserialized_packets = packets_to_process.iter().map(|p| &**p);
-        let (transactions, transaction_to_packet_indexes): (Vec<SanitizedTransaction>, Vec<usize>) =
-            deserialized_packets
-                .enumerate()
-                .filter_map(|(packet_index, deserialized_packet)| {
-                    if !buffered_packet_batches.is_forwarded(deserialized_packet) {
-                        unprocessed_packet_batches::transaction_from_deserialized_packet(
-                            deserialized_packet,
-                            &bank.feature_set,
-                            bank.vote_only_bank(),
-                            bank.as_ref(),
-                        )
-                        .map(|transaction| (transaction, packet_index))
-                    } else {
-                        None
-                    }
-                })
-                .unzip();
-
-        // report metrics
-        inc_new_counter_info!("banking_stage-packet_conversion", 1);
-        let unsanitized_packets_filtered_count =
-            packets_to_process.len().saturating_sub(transactions.len());
-        inc_new_counter_info!(
-            "banking_stage-dropped_tx_before_forwarding",
-            unsanitized_packets_filtered_count
-        );
-
-        (transactions, transaction_to_packet_indexes)
-    }
-
-    /// Checks sanitized transactions against bank, returns valid transaction indexes
-    fn filter_invalid_transactions(
-        transactions: &[SanitizedTransaction],
-        bank: &Arc<Bank>,
-    ) -> Vec<usize> {
-        let filter = vec![Ok(()); transactions.len()];
-        let results = Self::bank_check_transactions(bank, transactions, &filter);
-        // report metrics
-        let filtered_out_transactions_count = transactions.len().saturating_sub(results.len());
-        inc_new_counter_info!(
-            "banking_stage-dropped_tx_before_forwarding",
-            filtered_out_transactions_count
-        );
-
-        results
-            .iter()
-            .enumerate()
-            .filter_map(
-                |(tx_index, (result, _))| if result.is_ok() { Some(tx_index) } else { None },
-            )
-            .collect_vec()
-    }
-
-    fn prepare_filtered_packet_indexes(
-        transaction_to_packet_indexes: &[usize],
-        retained_transaction_indexes: &[usize],
-    ) -> Vec<usize> {
-        retained_transaction_indexes
-            .iter()
-            .map(|tx_index| transaction_to_packet_indexes[*tx_index])
-            .collect_vec()
-    }
-
-    fn collect_retained_packets(
-        buffered_packet_batches: &mut UnprocessedPacketBatches,
-        packets_to_process: &[Arc<ImmutableDeserializedPacket>],
-        retained_packet_indexes: &[usize],
-    ) -> Vec<Arc<ImmutableDeserializedPacket>> {
-        Self::remove_non_retained_packets(
-            buffered_packet_batches,
-            packets_to_process,
-            retained_packet_indexes,
-        );
-        retained_packet_indexes
-            .iter()
-            .map(|i| packets_to_process[*i].clone())
-            .collect_vec()
-    }
-
-    /// remove packets from UnprocessedPacketBatches.message_hash_to_transaction after they have
-    /// been removed from UnprocessedPacketBatches.packet_priority_queue
-    pub(crate) fn remove_non_retained_packets(
-        buffered_packet_batches: &mut UnprocessedPacketBatches,
-        packets_to_process: &[Arc<ImmutableDeserializedPacket>],
-        retained_packet_indexes: &[usize],
-    ) {
-        unprocessed_transaction_storage::filter_processed_packets(
-            retained_packet_indexes
-                .iter()
-                .chain(std::iter::once(&packets_to_process.len())),
-            |start, end| {
-                for processed_packet in &packets_to_process[start..end] {
-                    buffered_packet_batches
-                        .message_hash_to_transaction
-                        .remove(processed_packet.message_hash());
-                }
-            },
-        )
-    }
-
-    /// try to add filtered forwardable and valid packets to forward buffer;
-    /// returns vector of packet indexes that were accepted for forwarding.
-    fn add_filtered_packets_to_forward_buffer(
-        forward_buffer: &mut ForwardPacketBatchesByAccounts,
-        packets_to_process: &[Arc<ImmutableDeserializedPacket>],
-        transactions: &[SanitizedTransaction],
-        transaction_to_packet_indexes: &[usize],
-        forwardable_transaction_indexes: &[usize],
-        dropped_tx_before_forwarding_count: &mut usize,
-    ) -> Vec<usize> {
-        let mut added_packets_count: usize = 0;
-        let mut accepted_packet_indexes = Vec::with_capacity(transaction_to_packet_indexes.len());
-        for forwardable_transaction_index in forwardable_transaction_indexes {
-            let sanitized_transaction = &transactions[*forwardable_transaction_index];
-            let forwardable_packet_index =
-                transaction_to_packet_indexes[*forwardable_transaction_index];
-            let immutable_deserialized_packet =
-                packets_to_process[forwardable_packet_index].clone();
-            if !forward_buffer.try_add_packet(sanitized_transaction, immutable_deserialized_packet)
-            {
-                break;
-            }
-            accepted_packet_indexes.push(forwardable_packet_index);
-            saturating_add_assign!(added_packets_count, 1);
-        }
-
-        // count the packets not being forwarded in this batch
-        saturating_add_assign!(
-            *dropped_tx_before_forwarding_count,
-            forwardable_transaction_indexes.len() - added_packets_count
-        );
-
-        accepted_packet_indexes
     }
 
     #[allow(clippy::too_many_arguments)]
@@ -2096,35 +1803,6 @@ impl BankingStage {
             .collect_vec()
     }
 
-    /// Checks a batch of sanitized transactions again bank for age and status
-    fn bank_check_transactions(
-        bank: &Arc<Bank>,
-        transactions: &[SanitizedTransaction],
-        filter: &[transaction::Result<()>],
-    ) -> Vec<TransactionCheckResult> {
-        let mut error_counters = TransactionErrorMetrics::default();
-        // The following code also checks if the blockhash for a transaction is too old
-        // The check accounts for
-        //  1. Transaction forwarding delay
-        //  2. The slot at which the next leader will actually process the transaction
-        // Drop the transaction if it will expire by the time the next node receives and processes it
-        let api = perf_libs::api();
-        let max_tx_fwd_delay = if api.is_none() {
-            MAX_TRANSACTION_FORWARDING_DELAY
-        } else {
-            MAX_TRANSACTION_FORWARDING_DELAY_GPU
-        };
-
-        bank.check_transactions(
-            transactions,
-            filter,
-            (MAX_PROCESSING_AGE)
-                .saturating_sub(max_tx_fwd_delay)
-                .saturating_sub(FORWARD_TRANSACTIONS_TO_LEADER_AT_SLOT_OFFSET as usize),
-            &mut error_counters,
-        )
-    }
-
     /// This function filters pending packets that are still valid
     /// # Arguments
     /// * `transactions` - a batch of transactions deserialized from packets
@@ -2139,7 +1817,11 @@ impl BankingStage {
         let filter =
             Self::prepare_filter_for_pending_transactions(transactions.len(), pending_indexes);
 
-        let results = Self::bank_check_transactions(bank, transactions, &filter);
+        let results = bank.check_transactions_with_forwarding_delay(
+            transactions,
+            &filter,
+            FORWARD_TRANSACTIONS_TO_LEADER_AT_SLOT_OFFSET,
+        );
 
         Self::filter_valid_transaction_indexes(&results, transaction_to_packet_indexes)
     }
@@ -2165,13 +1847,13 @@ impl BankingStage {
             deserialized_packets
                 .enumerate()
                 .filter_map(|(i, deserialized_packet)| {
-                    unprocessed_packet_batches::transaction_from_deserialized_packet(
-                        deserialized_packet,
-                        &bank.feature_set,
-                        bank.vote_only_bank(),
-                        bank.as_ref(),
-                    )
-                    .map(|transaction| (transaction, i))
+                    deserialized_packet
+                        .compute_sanitized_transaction(
+                            &bank.feature_set,
+                            bank.vote_only_bank(),
+                            bank.as_ref(),
+                        )
+                        .map(|transaction| (transaction, i))
                 })
                 .unzip(),
             "packet_conversion",
@@ -2411,6 +2093,7 @@ where
 mod tests {
     use {
         super::*,
+        crate::unprocessed_packet_batches,
         crossbeam_channel::{unbounded, Receiver},
         solana_address_lookup_table_program::state::{AddressLookupTable, LookupTableMeta},
         solana_entry::entry::{next_entry, next_versioned_entry, Entry, EntrySlice},
@@ -2455,7 +2138,6 @@ mod tests {
             sync::atomic::{AtomicBool, Ordering},
             thread::sleep,
         },
-        unprocessed_packet_batches::DeserializedPacket,
     };
 
     fn new_test_cluster_info(contact_info: ContactInfo) -> ClusterInfo {
@@ -3485,144 +3167,6 @@ mod tests {
             assert!(commit_transactions_result.is_ok());
         }
         Blockstore::destroy(ledger_path.path()).unwrap();
-    }
-
-    #[test]
-    fn test_filter_and_forward_with_account_limits() {
-        solana_logger::setup();
-        let GenesisConfigInfo {
-            genesis_config,
-            mint_keypair,
-            ..
-        } = create_genesis_config(10);
-        let current_bank = Arc::new(Bank::new_for_tests(&genesis_config));
-
-        let simple_transactions: Vec<Transaction> = (0..256)
-            .map(|_id| {
-                // packets are deserialized upon receiving, failed packets will not be
-                // forwarded; Therefore we need to create real packets here.
-                let key1 = Keypair::new();
-                system_transaction::transfer(
-                    &mint_keypair,
-                    &key1.pubkey(),
-                    genesis_config.rent.minimum_balance(0),
-                    genesis_config.hash(),
-                )
-            })
-            .collect_vec();
-
-        let mut packets: Vec<DeserializedPacket> = simple_transactions
-            .iter()
-            .enumerate()
-            .map(|(packets_id, transaction)| {
-                let mut p = Packet::from_data(None, transaction).unwrap();
-                p.meta.port = packets_id as u16;
-                p.meta.set_tracer(true);
-                DeserializedPacket::new(p).unwrap()
-            })
-            .collect_vec();
-
-        // all packets are forwarded
-        {
-            let mut buffered_packet_batches: UnprocessedPacketBatches =
-                UnprocessedPacketBatches::from_iter(packets.clone().into_iter(), packets.len());
-            let mut forward_packet_batches_by_accounts =
-                ForwardPacketBatchesByAccounts::new_with_default_batch_limits();
-
-            let FilterForwardingResults {
-                total_forwardable_packets,
-                total_tracer_packets_in_buffer,
-                total_forwardable_tracer_packets,
-                ..
-            } = BankingStage::filter_and_forward_with_account_limits(
-                &current_bank,
-                &mut buffered_packet_batches,
-                &mut forward_packet_batches_by_accounts,
-                UNPROCESSED_BUFFER_STEP_SIZE,
-            );
-            assert_eq!(total_forwardable_packets, 256);
-            assert_eq!(total_tracer_packets_in_buffer, 256);
-            assert_eq!(total_forwardable_tracer_packets, 256);
-
-            // packets in a batch are forwarded in arbitrary order; verify the ports match after
-            // sorting
-            let expected_ports: Vec<_> = (0..256).collect();
-            let mut forwarded_ports: Vec<_> = forward_packet_batches_by_accounts
-                .iter_batches()
-                .flat_map(|batch| {
-                    batch
-                        .get_forwardable_packets()
-                        .into_iter()
-                        .map(|p| p.meta.port)
-                })
-                .collect();
-            forwarded_ports.sort_unstable();
-            assert_eq!(expected_ports, forwarded_ports);
-        }
-
-        // some packets are forwarded
-        {
-            let num_already_forwarded = 16;
-            for packet in &mut packets[0..num_already_forwarded] {
-                packet.forwarded = true;
-            }
-            let mut buffered_packet_batches: UnprocessedPacketBatches =
-                UnprocessedPacketBatches::from_iter(packets.clone().into_iter(), packets.len());
-            let mut forward_packet_batches_by_accounts =
-                ForwardPacketBatchesByAccounts::new_with_default_batch_limits();
-            let FilterForwardingResults {
-                total_forwardable_packets,
-                total_tracer_packets_in_buffer,
-                total_forwardable_tracer_packets,
-                ..
-            } = BankingStage::filter_and_forward_with_account_limits(
-                &current_bank,
-                &mut buffered_packet_batches,
-                &mut forward_packet_batches_by_accounts,
-                UNPROCESSED_BUFFER_STEP_SIZE,
-            );
-            assert_eq!(
-                total_forwardable_packets,
-                packets.len() - num_already_forwarded
-            );
-            assert_eq!(total_tracer_packets_in_buffer, packets.len());
-            assert_eq!(
-                total_forwardable_tracer_packets,
-                packets.len() - num_already_forwarded
-            );
-        }
-
-        // some packets are invalid (already processed)
-        {
-            let num_already_processed = 16;
-            for tx in &simple_transactions[0..num_already_processed] {
-                assert_eq!(current_bank.process_transaction(tx), Ok(()));
-            }
-            let mut buffered_packet_batches: UnprocessedPacketBatches =
-                UnprocessedPacketBatches::from_iter(packets.clone().into_iter(), packets.len());
-            let mut forward_packet_batches_by_accounts =
-                ForwardPacketBatchesByAccounts::new_with_default_batch_limits();
-            let FilterForwardingResults {
-                total_forwardable_packets,
-                total_tracer_packets_in_buffer,
-                total_forwardable_tracer_packets,
-                ..
-            } = BankingStage::filter_and_forward_with_account_limits(
-                &current_bank,
-                &mut buffered_packet_batches,
-                &mut forward_packet_batches_by_accounts,
-                UNPROCESSED_BUFFER_STEP_SIZE,
-            );
-            assert_eq!(
-                total_forwardable_packets,
-                packets.len() - num_already_processed
-            );
-            assert_eq!(total_tracer_packets_in_buffer, packets.len());
-            assert_eq!(
-                total_forwardable_tracer_packets,
-                packets.len() - num_already_processed
-            );
-        }
     }
 
     #[test]
