+++ core/src/forward_packet_batches_by_accounts.rs
@@ -182,7 +182,7 @@ impl ForwardPacketBatchesByAccounts {
 mod tests {
     use {
         super::*,
-        crate::unprocessed_packet_batches::{self, DeserializedPacket},
+        crate::unprocessed_packet_batches::DeserializedPacket,
         solana_runtime::transaction_priority_details::TransactionPriorityDetails,
         solana_sdk::{
             feature_set::FeatureSet, hash::Hash, signature::Keypair, system_transaction,
@@ -352,13 +352,14 @@ mod tests {
         // assert it is added, and buffer still accepts more packets
         {
             let packet = build_deserialized_packet_for_test(10, &hot_account, requested_cu);
-            let tx = unprocessed_packet_batches::transaction_from_deserialized_packet(
-                packet.immutable_section(),
-                &Arc::new(FeatureSet::default()),
-                false, //votes_only,
-                SimpleAddressLoader::Disabled,
-            )
-            .unwrap();
+            let tx = packet
+                .immutable_section()
+                .compute_sanitized_transaction(
+                    &Arc::new(FeatureSet::default()),
+                    false, //votes_only,
+                    SimpleAddressLoader::Disabled,
+                )
+                .unwrap();
             assert!(forward_packet_batches_by_accounts
                 .try_add_packet(&tx, packet.immutable_section().clone()));
 
@@ -372,13 +373,14 @@ mod tests {
         {
             let packet =
                 build_deserialized_packet_for_test(100, &hot_account, 1 /*requested_cu*/);
-            let tx = unprocessed_packet_batches::transaction_from_deserialized_packet(
-                packet.immutable_section(),
-                &Arc::new(FeatureSet::default()),
-                false, //votes_only,
-                SimpleAddressLoader::Disabled,
-            )
-            .unwrap();
+            let tx = packet
+                .immutable_section()
+                .compute_sanitized_transaction(
+                    &Arc::new(FeatureSet::default()),
+                    false, //votes_only,
+                    SimpleAddressLoader::Disabled,
+                )
+                .unwrap();
             assert!(!forward_packet_batches_by_accounts
                 .try_add_packet(&tx, packet.immutable_section().clone()));
 
@@ -392,13 +394,14 @@ mod tests {
         {
             let packet =
                 build_deserialized_packet_for_test(100, &other_account, 1 /*requested_cu*/);
-            let tx = unprocessed_packet_batches::transaction_from_deserialized_packet(
-                packet.immutable_section(),
-                &Arc::new(FeatureSet::default()),
-                false, //votes_only,
-                SimpleAddressLoader::Disabled,
-            )
-            .unwrap();
+            let tx = packet
+                .immutable_section()
+                .compute_sanitized_transaction(
+                    &Arc::new(FeatureSet::default()),
+                    false, //votes_only,
+                    SimpleAddressLoader::Disabled,
+                )
+                .unwrap();
             assert!(!forward_packet_batches_by_accounts
                 .try_add_packet(&tx, packet.immutable_section().clone()));
