+++ core/src/latest_unprocessed_votes.rs
@@ -2,7 +2,6 @@ use {
     crate::{
         forward_packet_batches_by_accounts::ForwardPacketBatchesByAccounts,
         immutable_deserialized_packet::{DeserializedPacketError, ImmutableDeserializedPacket},
-        unprocessed_packet_batches,
     },
     itertools::Itertools,
     rand::{thread_rng, Rng},
@@ -231,7 +230,7 @@ impl LatestUnprocessedVotes {
     /// Votes from validators with 0 stakes are ignored
     pub fn get_and_insert_forwardable_packets(
         &self,
-        bank: Arc<Bank>,
+        bank: &Arc<Bank>,
         forward_packet_batches_by_accounts: &mut ForwardPacketBatchesByAccounts,
     ) -> usize {
         let mut continue_forwarding = true;
@@ -250,9 +249,8 @@ impl LatestUnprocessedVotes {
                     let mut vote = lock.write().unwrap();
                     if !vote.is_vote_taken() && !vote.is_forwarded() {
                         let deserialized_vote_packet = vote.vote.as_ref().unwrap().clone();
-                        if let Some(sanitized_vote_transaction) =
-                            unprocessed_packet_batches::transaction_from_deserialized_packet(
-                                &deserialized_vote_packet,
+                        if let Some(sanitized_vote_transaction) = deserialized_vote_packet
+                            .compute_sanitized_transaction(
                                 &bank.feature_set,
                                 bank.vote_only_bank(),
                                 bank.as_ref(),
@@ -599,7 +597,7 @@ mod tests {
 
         // Don't forward 0 stake accounts
         let forwarded = latest_unprocessed_votes
-            .get_and_insert_forwardable_packets(bank, &mut forward_packet_batches_by_accounts);
+            .get_and_insert_forwardable_packets(&bank, &mut forward_packet_batches_by_accounts);
         assert_eq!(0, forwarded);
         assert_eq!(
             0,
@@ -621,7 +619,7 @@ mod tests {
 
         // Don't forward votes from gossip
         let forwarded = latest_unprocessed_votes.get_and_insert_forwardable_packets(
-            Arc::new(bank),
+            &Arc::new(bank),
             &mut forward_packet_batches_by_accounts,
         );
 
@@ -645,10 +643,8 @@ mod tests {
             ForwardPacketBatchesByAccounts::new_with_default_batch_limits();
 
         // Forward from TPU
-        let forwarded = latest_unprocessed_votes.get_and_insert_forwardable_packets(
-            bank.clone(),
-            &mut forward_packet_batches_by_accounts,
-        );
+        let forwarded = latest_unprocessed_votes
+            .get_and_insert_forwardable_packets(&bank, &mut forward_packet_batches_by_accounts);
 
         assert_eq!(1, forwarded);
         assert_eq!(
@@ -663,7 +659,7 @@ mod tests {
         let mut forward_packet_batches_by_accounts =
             ForwardPacketBatchesByAccounts::new_with_default_batch_limits();
         let forwarded = latest_unprocessed_votes
-            .get_and_insert_forwardable_packets(bank, &mut forward_packet_batches_by_accounts);
+            .get_and_insert_forwardable_packets(&bank, &mut forward_packet_batches_by_accounts);
 
         assert_eq!(0, forwarded);
         assert_eq!(
