+++ runtime/Cargo.toml
@@ -48,6 +48,7 @@ solana-frozen-abi = { path = "../frozen-abi", version = "=1.15.0" }
 solana-frozen-abi-macro = { path = "../frozen-abi/macro", version = "=1.15.0" }
 solana-measure = { path = "../measure", version = "=1.15.0" }
 solana-metrics = { path = "../metrics", version = "=1.15.0" }
+solana-perf = { path = "../perf", version = "=1.15.0" }
 solana-program-runtime = { path = "../program-runtime", version = "=1.15.0" }
 solana-rayon-threadlimit = { path = "../rayon-threadlimit", version = "=1.15.0" }
 solana-sdk = { path = "../sdk", version = "=1.15.0" }
